separated methods into classes, and classes into files
